1.small bugfix in test_wgraph_main

+++ inc/rb_tree.h
@@ -38,6 +38,6 @@ rbtree_node * rbtree_insert(rb_tree & tr, void * data);
 rbtree_node * rbtree_search(rb_tree const& tr, void const* data);
 
 void rbtree_inorder_walk(rb_tree const& tr);
-void rbtree_inorder_walk(rb_tree const& tr, void ** nodes, int& len);
+void rbtree_inorder_walk(rb_tree const& tr, void ** nodes, size_t& len);
 
 #endif /* HONGJUN_LIAO_RB_TREE_H_ */
